nanocoap: options_sort as pseudomodule
