+++ sys/net/application_layer/nanocoap/nanocoap.c
@@ -30,13 +30,13 @@
 #define ENABLE_DEBUG (0)
 #include "debug.h"
 
+static int _decode_value(unsigned val, uint8_t **pkt_pos_ptr, uint8_t *pkt_end);
 int coap_get_option_uint(coap_pkt_t *pkt, unsigned opt_num, uint32_t *target);
 static uint32_t _decode_uint(uint8_t *pkt_pos, unsigned nbytes);
 static size_t _encode_uint(uint32_t *val);
 
-extern int _decode_value(unsigned val, uint8_t **pkt_pos_ptr, uint8_t *pkt_end);
-extern uint8_t *_parse_option(coap_pkt_t *pkt, uint8_t *pkt_pos, uint16_t *delta, int *opt_len);
-extern ssize_t _sort_opts(coap_pkt_t *pkt);
+static uint8_t *_parse_option(coap_pkt_t *pkt, uint8_t *pkt_pos, uint16_t *delta, int *opt_len);
+static ssize_t _sort_opts(coap_pkt_t *pkt);
 
 /* http://tools.ietf.org/html/rfc7252#section-3
  *  0                   1                   2                   3
@@ -150,6 +150,22 @@ uint8_t *coap_find_option(coap_pkt_t *pkt, unsigned opt_num)
         optpos++;
     }
     return NULL;
+}
+
+static uint8_t *_parse_option(coap_pkt_t *pkt, uint8_t *pkt_pos, uint16_t *delta, int *opt_len)
+{
+    uint8_t *hdr_end = pkt->payload;
+
+    if (pkt_pos == hdr_end) {
+        return NULL;
+    }
+
+    uint8_t option_byte = *pkt_pos++;
+
+    *delta = _decode_value(option_byte >> 4, &pkt_pos, hdr_end);
+    *opt_len = _decode_value(option_byte & 0xf, &pkt_pos, hdr_end);
+
+    return pkt_pos;
 }
 
 int coap_get_option_uint(coap_pkt_t *pkt, unsigned opt_num, uint32_t *target)
@@ -381,6 +397,53 @@ void coap_pkt_init(coap_pkt_t *pkt, uint8_t *buf, size_t len, size_t header_len)
     pkt->payload_len = len - header_len;
 }
 
+static int _decode_value(unsigned val, uint8_t **pkt_pos_ptr, uint8_t *pkt_end)
+{
+    uint8_t *pkt_pos = *pkt_pos_ptr;
+    size_t left = pkt_end - pkt_pos;
+    int res;
+
+    switch (val) {
+        case 13:
+        {
+            /* An 8-bit unsigned integer follows the initial byte and
+               indicates the Option Delta minus 13. */
+            if (left < 1) {
+                return -ENOSPC;
+            }
+            uint8_t delta = *pkt_pos++;
+            res = delta + 13;
+            break;
+        }
+        case 14:
+        {
+            /* A 16-bit unsigned integer in network byte order follows
+             * the initial byte and indicates the Option Delta minus
+             * 269. */
+            if (left < 2) {
+                return -ENOSPC;
+            }
+            uint16_t delta;
+            uint8_t *_tmp = (uint8_t *)&delta;
+            *_tmp++ = *pkt_pos++;
+            *_tmp++ = *pkt_pos++;
+            res = ntohs(delta) + 269;
+            break;
+        }
+        case 15:
+            /* Reserved for the Payload Marker.  If the field is set to
+             * this value but the entire byte is not the payload
+             * marker, this MUST be processed as a message format
+             * error. */
+            return -EBADMSG;
+        default:
+            res = val;
+    }
+
+    *pkt_pos_ptr = pkt_pos;
+    return res;
+}
+
 static uint32_t _decode_uint(uint8_t *pkt_pos, unsigned nbytes)
 {
     assert(nbytes <= 4);
@@ -569,6 +632,79 @@ static ssize_t _add_opt_pkt(coap_pkt_t *pkt, uint16_t optnum, uint8_t *val,
 
     return optlen;
 }
+
+#ifdef MODULE_NANOCOAP_OPTIONS_SORT
+ssize_t _sort_opts(coap_pkt_t *pkt)
+{
+    if (pkt->options_len <= 1) {
+        return 0;
+    }
+
+    uint8_t *options = (uint8_t *)pkt->hdr + coap_get_total_hdr_len(pkt);
+    /* scratch buffer to hold sorted options, with extra space to handle any
+     * large option header */
+    uint8_t scratch[(pkt->payload - options) + 4];
+    uint8_t *scratch_pos = &scratch[0];
+
+    unsigned sorted   = 0;        /* count of sorted elements */
+    uint8_t *next_pos = options;
+
+    do {
+        // find next option to write
+        unsigned i, best_i = sorted;
+        for (i = sorted+1; i < pkt->options_len; i++) {
+            if (pkt->options[i].opt_num < pkt->options[best_i].opt_num) {
+                best_i = i;
+            }
+        }
+        i = best_i;
+        /* prepare to relocate option memo currently at sorted location */
+        coap_optpos_t bubble_opt;
+        bubble_opt.opt_num = pkt->options[sorted].opt_num;
+        bubble_opt.offset  = pkt->options[sorted].offset;
+
+        /* read selected option */
+        uint16_t delta;      // unused
+        int option_len;
+        uint8_t *val = _parse_option(pkt, (uint8_t *)pkt->hdr + pkt->options[i].offset,
+                                     &delta, &option_len);
+
+        /* write option to scratch buffer */
+        uint8_t *last_pos = next_pos;
+        next_pos += coap_put_option(scratch_pos,
+                                    sorted ? pkt->options[sorted-1].opt_num : 0,
+                                    pkt->options[i].opt_num, val, option_len);
+        scratch_pos += (next_pos - last_pos);
+
+        /* update pkt->options record and save bubbled option */
+        pkt->options[sorted].opt_num = pkt->options[i].opt_num;
+        pkt->options[sorted].offset  = last_pos - (uint8_t *)pkt->hdr;
+        if (i != sorted) {
+            pkt->options[i].opt_num = bubble_opt.opt_num;
+            pkt->options[i].offset  = bubble_opt.offset;
+        }
+    } while (++sorted < pkt->options_len);
+
+    int len_delta = next_pos - pkt->payload;
+    DEBUG("sort moves payload by %d\n", len_delta);
+
+    memcpy(options, &scratch[0], next_pos - options);
+
+    pkt->payload     += len_delta;
+    pkt->payload_len -= len_delta;
+    return 0;
+}
+
+bool coap_opt_sorted(coap_pkt_t *pkt)
+{
+    for (unsigned i = 1; i < pkt->options_len; i++) {
+        if (pkt->options[i].opt_num < pkt->options[i-1].opt_num) {
+            return false;
+        }
+    }
+    return true;
+}
+#endif
 
 ssize_t coap_opt_add_string(coap_pkt_t *pkt, uint16_t optnum, const char *string,
                            char separator)
